% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leanplumber.R
\docType{package}
\name{leanplumber-package}
\alias{leanplumber}
\alias{leanplumber-package}
\title{Thin Wrapper for the Leanplum API}
\description{
The leanplumber package provides a simple interface in R to Leanplum's API.
}
\details{
Start with the documentation at https://docs.leanplum.com/reference#introduction.
This package makes it a bit easier to send data to Leanplum, as well as export data.
Leanplum API endpoints and functions currently supported include:
exportData - Exports All Data to CSV or JSON
exportReport - Exports statistics for user activity, messages, and A/B tests
exportUser - Get information about a specific user
exportUsers - Get a list of all userIds that match a condition
setUserAttributes - Insert and/or update unique attributes as a key-value map
getMessages - Get a list of all active messages
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/raybuhr/leanplumber}
  \item Report bugs at \url{https://github.com/raybuhr/leanplumber/issues}
}

}
\author{
\strong{Maintainer}: Ray Buhr \email{raymond.buhr@gmail.com}

}
\keyword{internal}
