% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{lp_export_data}
\alias{lp_export_data}
\title{Export All Data to CSV or JSON}
\usage{
lp_export_data(start_date = Sys.Date() - 1, end_date = NULL,
  export_format = NULL, api_version = "1.0.6",
  app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_DATA_EXPORT"))
}
\arguments{
\item{start_date}{string - First date in range to include in PDT/PST format: YYYYmmdd. Example: 20140223.}

\item{end_date}{string - Last date in range to include in PDT/PST format: YYYYmmdd. Defaults to startDate if not provided. Example: 20140223.}

\item{export_format}{string - The format to export data. Can be either json or csv. Default: json.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Data Export key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#get_api-action-exportdata
}
\details{
Exports raw data to downloadable files. Data is split into roughly 256 MB files, and is not necessarily ordered.
Exports can be made in JSON or CSV format
For JSON format, each file contains 1 line per session, with each session JSON-encoded.
For CSV format, data is split into separate files for sessions, states, events, event parameters, and user attributes.
}
