% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_message_report.R
\name{lp_export_message_results}
\alias{lp_export_message_results}
\title{Export Message Results Report}
\usage{
lp_export_message_results(message_id, event_names = NULL,
  start_date = Sys.Date() - 1, end_date = NULL, api_version = "1.0.6",
  app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_DATA_EXPORT"))
}
\arguments{
\item{message_id}{number - the message_id, can be found by clicking on the message page, e.g. 5098409547661312}

\item{event_names}{character vector of the events to be included in the report. Example: c\\("Add to cart", "Purchase"\\)}

\item{start_date}{string - First date in range to include in PDT/PST format: YYYYmmdd. Example: 20140223.}

\item{end_date}{string - Last date in range to include in PDT/PST format: YYYYmmdd. Defaults to startDate if not provided. Example: 20140223.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Data Export key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#get_api-action-exportreport
}
\details{
Exports statistics for a message from your app just as in the Analytics tab of the dashboard
Report are over specified period of time.
Report data becomes available to export every 2 hours, and only for complete sessions.
You may only export report data 100 times per day per app.
Exports with invalid arguments do not count towards the limit.
}
