% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_users.R
\name{lp_export_users}
\alias{lp_export_users}
\title{Export Users exports multiple user IDs}
\usage{
lp_export_users(segment = NULL, ab_test_id = NULL, api_version = "1.0.6",
  app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_DATA_EXPORT"))
}
\arguments{
\item{segment}{string - Limit export to only users that match the segment. The syntax is identical to that produced by the "Insert Value" feature on the dashboard.}

\item{ab_test_id}{integer - Limit export to only users that are in the given A/B test. The output will include the variant ID and name for each user.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Data Export key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#get_api-action-exportusers
}
\details{
Exports multiple user IDs.
This may be executed up to 10 times successfully per day.
This method requires your data export API clientKey.
Use getExportResults with the returned jobId to get the results.
}
