% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{lp_get_export_results}
\alias{lp_get_export_results}
\title{Export results retrieves the result of an export job generated by: exportData, exportReport, and exportUsers.}
\usage{
lp_get_export_results(job_id, api_version = "1.0.6",
  app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_DATA_EXPORT"))
}
\arguments{
\item{job_id}{string - The export job ID.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Data Export key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#get_api-action-exportusers
}
\details{
Retrieves the result of an export job generated by:
exportData, exportReport, and exportUsers.
This method requires your data export API clientKey.
}
