% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{lp_get_messages}
\alias{lp_get_messages}
\title{Gets information about all the messages for a given app}
\usage{
lp_get_messages(include_drafts = "false", recent = "true",
  api_version = "1.0.6", app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_READ_CONTENT"))
}
\arguments{
\item{include_drafts}{string - Include drafts and unpublished changes. Default: false.}

\item{recent}{string - Only return information about active or recently finished messages. Default: true.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Read-only Content key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#get_api-action-getmessages
}
\details{
A successful response will return a list of messages.
}
