% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_user.R
\name{lp_get_user_info}
\alias{lp_get_user_info}
\title{Export user retrieves attributes for a single user}
\usage{
lp_get_user_info(user_id, device_id = NULL, api_version = "1.0.6",
  app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_DATA_EXPORT"))
}
\arguments{
\item{user_id}{string - The current user ID. You can set this to an email address or whatever you use at your company for user IDs. Leave it blank to use the device ID. For more info, see selecting a user.}

\item{device_id}{string - A unique ID for the device targeted by the request. You must provide a deviceId and/or a userId.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Data Export key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#get_api-action-exportuser
}
\details{
A successful request will return a map of the user's attributes, events and states.
}
