% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_user_attributes.R
\name{lp_update_user_attributes}
\alias{lp_update_user_attributes}
\title{Sets user attributes for the user given by userId and/or deviceId.}
\usage{
lp_update_user_attributes(user_id, user_attributes, api_version = "1.0.6",
  app_id = Sys.getenv("LEANPLUM_APP_ID"),
  client_key = Sys.getenv("LEANPLUM_PRODUCTION"))
}
\arguments{
\item{user_id}{string - The current user ID. You can set this to an email address or whatever you use at your company for user IDs. Leave it blank to use the device ID. For more info, see selecting a user.}

\item{user_attributes}{object - A map of user attributes as key-value pairs. Each key must be a string. Attributes are saved across sessions. Only supplied attributes will be updated i.e., if you omit an existing attribute, it will be preserved. Example: \\{"gender":"F","age":21\\}.}

\item{api_version}{string - The version of the Leanplum API to use. The current version is 1.0.6.}

\item{app_id}{string - The application ID. To find yours, select your app in the navigation column, and click Edit Apps. Under Keys, click Show.}

\item{client_key}{string - The Production key for your Leanplum App.}
}
\value{
A http response parsed by httr and jsonlite
}
\description{
https://docs.leanplum.com/reference#post_api-action-setuserattributes
}
\details{
Sets user attributes for the user given by userId and/or deviceId.
If the user has an open session, the attributes for the current session will also be updated.
Attributes will then propagate on data going forward.
User properties not supplied in this method will not be affected.
If the user/device does not exist, a new user will be created.
}
